Add support for adding custom hashtags to the tweet
This commit adds support to define custom hashtags to add them
to the generated tweet. The github actions cannot take a list
as an input. So, the hashtags should be parsed as a string and
it is converted into list before adding them to the tweet. The
documentation is also updated.